set raredisease config style as default by nf-analysis
